+++ Source/AliveLibAO/TimerTrigger.hpp
@@ -36,7 +36,7 @@ public:
     Guid mTlvInfo;
     s32 mActivationDelayTimer = 0;
     s32 mActivationDelay = 0;
-    s16 mStartingSwitchState = 0;
+    s32 mStartingSwitchState = 0;
 };
